それに伴ってrspecを修正
