+++ app/views/admin/users/create.html.erb
@@ -7,7 +7,7 @@
         <div class="flex items-center justify-between">
           <h2 class="text-2xl font-bold text-gray-900">新規ユーザー登録</h2>
           <div class="flex items-center space-x-4">
-            <%= link_to admin_create_user_import_path, class: "bg-green-600 text-white px-4 py-2 rounded-md hover:bg-green-700 transition-colors duration-300" do %>
+            <%= link_to admin_users_import_path, class: "bg-green-600 text-white px-4 py-2 rounded-md hover:bg-green-700 transition-colors duration-300" do %>
               ユーザー一括登録
             <% end %>
             <%= link_to "管理者画面に戻る", admin_path, class: "text-sm font-medium text-blue-600 hover:text-blue-500" %>
