+++ app/views/admin/users/import.html.erb
@@ -47,7 +47,7 @@
                 <p class="text-sm text-gray-700 mb-2">まずはテンプレートファイルをダウンロードしてください</p>
                 <p class="text-xs text-gray-500">名前、メールアドレス、パスワード、ユーザータイプ、期数、性別、生年月日の項目があります</p>
               </div>
-              <%= link_to admin_create_user_import_template_path, class: "bg-blue-600 text-white px-4 py-2 rounded-md hover:bg-blue-700 transition-colors duration-300" do %>
+              <%= link_to admin_users_import_template_path, class: "bg-blue-600 text-white px-4 py-2 rounded-md hover:bg-blue-700 transition-colors duration-300" do %>
                 <svg class="inline-block w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                   <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 10v6m0 0l-3-3m3 3l3-3m2 8H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"></path>
                 </svg>
@@ -60,7 +60,7 @@
         <!-- Excelファイルアップロード -->
         <div class="mb-8">
           <h3 class="text-lg font-medium text-gray-900 mb-4">2. Excelファイルをアップロード</h3>
-          <%= form_with url: admin_create_user_import_preview_path, multipart: true, class: "space-y-4", id: "excel_upload_form" do |f| %>
+          <%= form_with url: admin_users_import_preview_path, multipart: true, class: "space-y-4", id: "excel_upload_form" do |f| %>
             <label for="excel_file" class="cursor-pointer block">
               <div class="border-2 border-dashed border-gray-300 rounded-lg p-8 hover:border-gray-400 transition-colors">
                 <div class="text-center">
@@ -97,7 +97,7 @@
               <div class="mb-4">
                 <div class="flex items-center justify-between">
                   <p class="text-sm font-medium text-gray-900"><%= @preview_data.size %>人のユーザーが登録されます</p>
-                  <%= link_to admin_create_user_import_path(clear_preview: true), class: "text-sm text-blue-600 hover:text-blue-500" do %>
+                  <%= link_to admin_users_import_path(clear_preview: true), class: "text-sm text-blue-600 hover:text-blue-500" do %>
                     プレビューをクリア
                   <% end %>
                 </div>
@@ -168,7 +168,7 @@
             キャンセル
           <% end %>
           <% if @preview_data.present? %>
-            <%= form_with url: admin_create_user_import_execute_path, method: :post, local: true, class: "inline-block" do |f| %>
+            <%= form_with url: admin_users_import_execute_path, method: :post, local: true, class: "inline-block" do |f| %>
               <%= f.submit "一括登録実行", class: "px-6 py-3 text-sm font-medium text-white bg-green-600 border border-transparent rounded-md hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-green-500", data: { confirm: "#{@preview_data.size}人のユーザーを一括登録しますか？" } %>
             <% end %>
           <% end %>
